+++ src/tint/transform/transform.cc
@@ -110,7 +110,7 @@ const ast::Type* Transform::CreateASTTypeFor(CloneContext& ctx, const sem::Type*
         if (a->IsRuntimeSized()) {
             return ctx.dst->ty.array(el, nullptr, std::move(attrs));
         } else {
-            return ctx.dst->ty.array(el, a->Count(), std::move(attrs));
+            return ctx.dst->ty.array(el, u32(a->Count()), std::move(attrs));
         }
     }
     if (auto* s = ty->As<sem::Struct>()) {
